SRC Updates
+ Updated Src Folder
+ Optimized Code
+ Optimized Types
