+++ .env
@@ -33,7 +33,7 @@ DEV_DB_NAME = FastifyTemplate_dev
 
 # Dev Redis
 DEV_REDIS_HOST= localhost
-DEV_REDIS_PORT = 6380
+DEV_REDIS_PORT = 6379
 
 # JWT Secret
 JWT_SECRET=0SK4zL2sdXqilQHxGB8Kb2UoGq7ZUF5PJaR44DCevb/sGXqN2pTlb9oQ92KJZpqUCQqqbDsH+oOqfTYhWoJola9dsSHdIGOeX65hNTH91lUFJo+liQCS2kUXCThdv1fuwVY3v+VifX9LTbYnBKQ/18iAGOwguz+5HHJrE7wBSMSmhrPxlvnOGnv0k5C79z48I1zTAzP931Qfs5wbk95IAuJ0zvk854uvHwFKzQ==
