+++ src/routes/websocket.ts
@@ -1,13 +1,13 @@
 import { Socket } from "socket.io";
-import Server from "../index";
+import { server } from "../lib/server";
 
 function SocketHandler(socket: Socket) {
   socket.on("connection", () => {
-    Server.log.info("Client connected", socket.id);
+    server.log.info("Client connected", socket.id);
   });
 
   socket.on("disconnect", () => {
-    Server.log.info("Client disconnected", socket.id);
+    server.log.info("Client disconnected", socket.id);
   });
 }
