+++ src/types/global.d.ts
@@ -1,0 +1,9 @@
+// ! Type definitions for global objects
+declare global {
+  interface IUserBody {
+    username: string;
+    password: string;
+  }
+}
+
+export {};
